% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{assess_performance}
\alias{assess_performance}
\title{Assess multi-block sPLSDA (DIABLO) performance}
\usage{
assess_performance(data, dist, ncomp)
}
\arguments{
\item{data}{list of dataframes with quantitative multi-omics data.}

\item{dist}{string describing distance metric "centroids.dist", "max.dist", "mahalanobis.dist". Defaults to "centroids.dist"}

\item{ncomp}{integer assigning number of components for multi-block sPLSDA. Defaults to 0.}
}
\description{
Assess multi-block sPLSDA (DIABLO) performance
}
\keyword{Internal}
