% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{classify_plsda}
\alias{classify_plsda}
\title{Perform PLSDA}
\usage{
classify_plsda(
  data,
  classes,
  pch = NA,
  title = "",
  ncomp = 0,
  contrib = "max",
  outdir = "./",
  mappings = NULL,
  dist = "centroids.dist",
  bg = TRUE
)
}
\arguments{
\item{data}{list of dataframes with original data}

\item{classes}{1-column dataframe with classes}

\item{pch}{1-column dataframe with repeated measurments. Defaults to NA.}

\item{title}{string to name plots with. Defaults to "".}

\item{ncomp}{integer assigning number of components for PLSDA. Defaults to 0.}

\item{contrib}{string to determine loading weights "max", "min". Defaults to "max".}

\item{outdir}{string to outfile directory. Defaults to "./"}

\item{mappings}{dataframe of mappings for features. Defaults to NULL}

\item{dist}{string describing distance metric "centroids.dist", "max.dist", "mahalanobis.dist". Defaults to "centroids.dist"}

\item{bg}{boolean describing if background should be plotted. Defaults to TRUE}
}
\description{
Perform PLSDA (can account for repeated measurments if specified):
}
\seealso{
\code{\link[=classify_plsda]{classify_plsda()}}, \code{\link[=classify_splsda]{classify_splsda()}}
}
