% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{classify_splsda}
\alias{classify_splsda}
\title{Perform sPLSDA}
\usage{
classify_splsda(
  data,
  classes,
  pch = NA,
  title = "",
  ncomp = NULL,
  keepX = NULL,
  contrib = "max",
  outdir = "./",
  mappings = NULL,
  dist = "centroids.dist",
  bg = TRUE,
  validation = "loo",
  nrepeat = 10,
  folds = 10,
  near_zero_var = FALSE
)
}
\arguments{
\item{data}{list of dataframes with original data}

\item{classes}{1-column dataframe with classes}

\item{pch}{1-column dataframe with repeated measurments. Defaults to NA.}

\item{title}{string to name plots with. Defaults to "".}

\item{ncomp}{integer assigning number of components for PLSDA. Defaults to 0.}

\item{keepX}{vector of integers for feature selection. Defaults to NULL.}

\item{contrib}{string to determine loading weights "max", "min". Defaults to "max".}

\item{outdir}{string to outfile directory. Defaults to "./"}

\item{mappings}{dataframe of mappings for features. Defaults to NULL}

\item{dist}{string describing distance metric "centroids.dist", "max.dist", "mahalanobis.dist". Defaults to "centroids.dist"}

\item{bg}{boolean describing if background should be plotted. Defaults to TRUE}

\item{validation}{character specifying "loo" or "M-fold" cross-validation. Defaults to "loo".}

\item{nrepeat}{if M-fold validation, number of repeats. Defaults to 10. No effect for "loo"}

\item{folds}{if M-fold validation, number of folds. Defaults to 10. No effect for "loo"}

\item{near_zero_var}{if variance approaches zero, enable this option}
}
\description{
Perform sPLSDA (can account for repeated measurments if specified):
}
\seealso{
\code{\link[=classify_plsda]{classify_plsda()}}, \code{\link[=classify_splsda]{classify_splsda()}}, \code{\link[=tune_splsda]{tune_splsda()}}
}
