% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{cor_imputed_unimputed}
\alias{cor_imputed_unimputed}
\title{Plot correlation between unimputed and imputed data}
\usage{
cor_imputed_unimputed(pca_withna, pca_impute, names)
}
\arguments{
\item{pca_withna}{list of dataframes with original data}

\item{pca_impute}{list of dataframes with imputed data}

\item{names}{vector of names for each data category}
}
\description{
Quality control plot to check if data has mutated significantly
}
\seealso{
\code{\link[=impute_missing]{impute_missing()}}, \code{\link[=replace_missing]{replace_missing()}}, \code{\link[=cor_imputed_unimputed]{cor_imputed_unimputed()}}
}
