% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{count_missing}
\alias{count_missing}
\title{Count missing values}
\usage{
count_missing(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Count missing values: list of named dataframes -> list of dataframes
}
