% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{create_design}
\alias{create_design}
\title{Create the design matrix}
\usage{
create_design(data, link = 0.1)
}
\arguments{
\item{data}{list of named dataframes}

\item{link}{linkage, (discrimination) 0 <-> 1 (correlation). Defaults to 0.1.}
}
\description{
Create the design matrix: list of named dataframes -> matrix
If your NA values are represented as 0 or something else, convert them first.
}
