% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{export_parameters}
\alias{export_parameters}
\title{Export tuned parameters}
\usage{
export_parameters(
  dist_plsda,
  plsda_ncomp,
  dist_splsda,
  splsda_ncomp,
  splsda_keepx,
  dist_diablo,
  diablo_ncomp,
  diablo_keepx,
  outdir = "./"
)
}
\arguments{
\item{dist_plsda}{list of distance metrics.}

\item{plsda_ncomp}{list of optimal components.}

\item{dist_splsda}{list of distance metrics.}

\item{splsda_ncomp}{list of optimal components.}

\item{splsda_keepx}{list of selected values from grid.}

\item{dist_diablo}{list of distance metrics.}

\item{diablo_ncomp}{list of optimal components.}

\item{diablo_keepx}{list of selected values from grid.}

\item{outfile_path}{path to tab separated output file.}
}
\description{
Export tuned parameters into a mapping table, format: dataframe -> outfile_path
These are generated automatically as part of the tuning process.
Can be loaded back in directly for tuning in the main script with --optimal_params
Note that the order of data blocks/omics must match the order provided in input --data and --data_names!
Also returns the dataframe.
}
\seealso{
\code{\link[=parse_parameters]{parse_parameters()}}
}
