% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{force_unique_blocks}
\alias{force_unique_blocks}
\title{Force unique blocks}
\usage{
force_unique_blocks(data)
}
\arguments{
\item{data}{dataframe with original data.}
}
\description{
Force each feature in each block of omics data to have unique names
}
