% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getJointRDA.R
\name{getJointRDA}
\alias{getJointRDA}
\alias{getJointRDA,MultiAssayExperiment-method}
\alias{getJointRDA,SingleCellExperiment-method}
\alias{getJointRDA,ANY-method}
\alias{plot.JointRDA}
\title{Redundancy analysis between two omic layers}
\usage{
getJointRDA(x, ...)

\S4method{getJointRDA}{MultiAssayExperiment}(x, experiments, assay.types, ...)

\S4method{getJointRDA}{SingleCellExperiment}(x, experiments, assay.types, ...)

\S4method{getJointRDA}{ANY}(x, ...)

\method{plot}{JointRDA}(x, show.species = TRUE, show.biplot = TRUE, ...)
}
\arguments{
\item{x}{\code{MultiAssayExperiment} or \code{SingleCellExperiment}.}

\item{...}{additional arguments passed to \code{vegan::rda()}.}

\item{experiments}{\code{Character vector} or \code{integer vector}. Names
or indices of experiments selected from \code{x}.}

\item{assay.types}{\code{Character vector}. Names of assays selected from
\code{experiments}. Must match the length of
\code{experiments}.}
}
\value{
An object of class \code{rda} as returned by \code{vegan::rda()}, containing
ordination results, explained variance, and model statistics.
}
\description{
Performs redundancy analysis (RDA) to assess linear associations between two
datasets (e.g., microbiome composition and metabolite profiles). The method
evaluates how much variation in one dataset can be explained by the other.
}
\details{
The function extracts matched samples from the selected experiments and
assays, aligns them, and applies redundancy analysis using
\code{vegan::rda()}.

RDA is a constrained ordination technique equivalent to multivariate linear
regression followed by principal component analysis (PCA) on fitted values.
}
\examples{
library(multiomics)
library(mia)

data(HintikkaXOData)
mae <- HintikkaXOData
# Replace missing values with 0
assay(mae[[2]])[is.na(assay(mae[[2]]))] <- 0

# Perform RDA
getJointRDA(
    mae,
    experiments = c(1, 2),
    assay.types = c("counts", "nmr")
)

}
\references{
Legendre, P., & Legendre, L. (2012). \emph{Numerical Ecology}.
Elsevier.
}
\seealso{
\code{\link[vegan:rda]{vegan::rda()}}
}
