% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getLASSO.R
\name{getLASSO}
\alias{getLASSO}
\alias{getLASSO,MultiAssayExperiment-method}
\alias{getLASSO,SingleCellExperiment-method}
\alias{getLASSO,ANY-method}
\alias{plot.LASSO}
\title{Sparse association modeling between microbiome and metabolites using LASSO}
\usage{
getLASSO(x, ...)

\S4method{getLASSO}{MultiAssayExperiment}(x, experiments, assay.types, ...)

\S4method{getLASSO}{SingleCellExperiment}(x, experiments, assay.types, ...)

\S4method{getLASSO}{ANY}(x, ...)

\method{plot}{LASSO}(x, sort = TRUE, ...)
}
\arguments{
\item{x}{\code{MultiAssayExperiment} or \code{SingleCellExperiment}.}

\item{...}{additional arguments passed to \code{glmnet::cv.glmnet()}.
\itemize{
\item \code{apply.logratio}: \code{Logical scalar}. Whether to apply
log-ratio transformation. (Default: \code{TRUE}).
}}

\item{experiments}{\code{Character vector} or \code{integer vector}. Names
or indices of experiments selected from \code{x}.}

\item{assay.types}{\code{Character vector}. Names of assays selected from
\code{experiments}. Must match the length of
\code{experiments}.}
}
\value{
A numeric matrix of class \code{matrix} with:
\itemize{
\item Rows: taxa (predictors)
\item Columns: metabolites (responses)
\item Values: LASSO regression coefficients
}
}
\description{
Performs LASSO regression to identify associations between microbiome
features (predictors) and metabolites (responses). For each metabolite, a
separate LASSO model is fitted to select the most relevant taxa and estimate
their regression coefficients.
}
\details{
The function fits a separate LASSO model for each metabolite using the
microbiome data as predictors. The procedure is implemented using
\code{glmnet} with cross-validation to select the optimal regularization
parameter.

Non-selected taxa (i.e., coefficients shrunk to zero) are represented as
\code{NA}.

The resulting matrix includes only core features, i.e., features including
only \code{NA}s are filtered out.

When applied to log-ratio transformed data (i.e.,
\code{apply.logratio=TRUE}), this approach is equivalent to
implementation in \code{coda4microbiome::coda_glmnet()}.
}
\examples{
library(multiomics)
library(mia)

data(HintikkaXOData)
mae <- HintikkaXOData
# Replace missing values with 0
assay(mae[[2]])[is.na(assay(mae[[2]]))] <- 0

# Agglomerate microbiome data
mae[[1]] <- agglomerateByRank(mae[[1]], rank = "Phylum")

# Perform LASSO
res <- getLASSO(
    mae,
    experiments = c(1, 2),
    assay.types = c("counts", "nmr")
)

# Visualize results
plotLASSO(res)

}
\references{
Calle M.L., Pujolassos, M. and Susin A. (2023).
coda4microbiome: compositional data analysis for microbiome cross-sectional
and longitudinal studies. \emph{BMC Bioinformatics}, 24, 82.
https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-023-05205-3
}
\seealso{
\code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
}
