% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getMantel.R
\name{getMantel}
\alias{getMantel}
\alias{getMantel,MultiAssayExperiment-method}
\alias{getMantel,SingleCellExperiment-method}
\alias{getMantel,ANY-method}
\title{Mantel test between two tables}
\usage{
getMantel(x, ...)

\S4method{getMantel}{MultiAssayExperiment}(x, experiments, assay.types, ...)

\S4method{getMantel}{SingleCellExperiment}(x, experiments, assay.types, ...)

\S4method{getMantel}{ANY}(x, dist.methods = c("bray", "euclidean"), group.by = NULL, ...)
}
\arguments{
\item{x}{\code{MultiAssayExperiment} or \code{SingleCellExperiment}.}

\item{...}{additional arguments:
\itemize{
\item \code{npermutations}: \code{Integer scalar}. Number of permutations
used to assess significance. (Default: \code{999}).
}}

\item{experiments}{\code{Character vector} or \code{integer vector}. Names
or indices of experiments selected from \code{x}.}

\item{assay.types}{\code{Character vector}. Names of assays selected from
\code{experiments}. Must match the length of
\code{experiments}.}

\item{dist.methods}{\code{Character vector}. Distance methods used to compute
dissimilarity matrices (one per dataset).  Must match the length of
\code{experiments}. Must be values from \code{vegan::vegdist()}.
(Default: \code{c("bray", "euclidean")})}

\item{group.by}{\code{Character scalar}. Column name in sample metadata used
for stratified permutations. If \code{NULL}, non-stratified test is
performed. (Default: \code{NULL})}

\item{method}{\code{Character scalar}. Correlation method for Mantel test.
Options include \code{"pearson"}, \code{"spearman"}, or \code{"kendall"}.
(Default: \code{"kendall"})}
}
\value{
An object of class \code{"mantel"} as returned by
\code{vegan::mantel()}.
}
\description{
Performs a Mantel test to assess the association between two data matrices
(e.g., different omics layers) by comparing their dissimilarity matrices.
Mantel test can be used to test whether there is general association between
two omics layers.
}
\details{
The Mantel test evaluates the correlation between two distance matrices.
For multi-assay objects, the function:
\enumerate{
\item extracts shared samples across selected experiments,
\item computes dissimilarity matrices for each dataset,
\item optionally applies stratification (group-wise permutations),
\item performs the Mantel test using \code{vegan::mantel()}.
}
}
\examples{
library(multiomics)
library(mia)

data(HintikkaXOData)
mae <- HintikkaXOData
# Replace missing values with 0
assay(mae[[2]])[is.na(assay(mae[[2]]))] <- 0

# Perform Mantel test
getMantel(
    mae,
    experiments = c(1, 2),
    assay.types = c("counts", "nmr")
)

}
\references{
Borcard, D. & Legendre, P. (2012) Is the Mantel correlogram powerful enough
to be useful in ecological analysis? A simulation study. Ecology 93:
1473-1481.
}
\seealso{
\code{\link[vegan:vegdist]{vegan::vegdist()}}
}
