% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing values}
\usage{
impute_missing(data, ncomps, outdir)
}
\arguments{
\item{data}{dataframe}

\item{ncomps}{number of components}

\item{outdir}{write imputed data to outfile path}
}
\description{
Impute missing values: list of named dataframes with NA -> list of dataframes
}
\seealso{
\code{\link[=impute_missing]{impute_missing()}}, \code{\link[=replace_missing]{replace_missing()}}, \code{\link[=cor_imputed_unimputed]{cor_imputed_unimputed()}}
}
