% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{parse_classes}
\alias{parse_classes}
\title{Load in class data}
\usage{
parse_classes(infile_path)
}
\arguments{
\item{infile_path}{path to tab separated input classes file with header}
}
\description{
Load in class data for data: infile_path -> vector (of strings)
}
