% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{parse_data}
\alias{parse_data}
\title{Load and parse data}
\usage{
parse_data(infile_path, offset = 0, missing_as = NA, rmna = TRUE)
}
\arguments{
\item{infile_path}{path to tab separated input file.}

\item{offset}{add offset to data. Defaults to 0 (no offset).}

\item{missing_as}{set missing values to this. Defaults to NA.}

\item{rmna}{remove NA values if a feature has only NA. Defaults to TRUE.}
}
\description{
Load in omics data into a diablo-compatible format: infile_path -> dataframe
}
