% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{parse_mappings}
\alias{parse_mappings}
\title{Load and parse mappings}
\usage{
parse_mappings(infile_path)
}
\arguments{
\item{infile_path}{path to tab separated input file.}
}
\description{
Load in id:name data into a mapping table, format: infile_path -> dataframe
Feature names can only reach a certain length and must be unique.
So you can use this if you want to remap identifiers out to something else.
}
\seealso{
\code{\link[=remap_data]{remap_data()}}
}
