% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{parse_parameters}
\alias{parse_parameters}
\title{Load and parse tuned parameters}
\usage{
parse_parameters(infile_path)
}
\arguments{
\item{infile_path}{path to tab separated input file.}
}
\description{
Load in tuned parameters into a mapping table, format: infile_path -> dataframe
Must follow specific format!
method           blocks       distance                                   ncomp keepx
<s/plsda/diablo> \if{html}{\out{<data block>}} <max.dist/centroids.dist/mahalanobis.dist> \if{html}{\out{<int>}} <1,2,3...>
...
}
\details{
For example:

method blocks       distance ncomp    keepx
plsda  lncrna centroids.dist     3        0
plsda   mirna centroids.dist     2        0
plsda    mrna centroids.dist     5        0
splsda lncrna centroids.dist     2     5,25
splsda  mirna centroids.dist     2   75,100
splsda   mrna centroids.dist     2    85,20
diablo lncrna centroids.dist     1        5
diablo  mirna centroids.dist     4 5,10,5,5
diablo   mrna centroids.dist     1       25
}
\seealso{
\code{\link[=export_parameters]{export_parameters()}}
}
