% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{plot_additional}
\alias{plot_additional}
\title{Plot correlation circle plots and biplots}
\usage{
plot_additional(data, data_pca, classes, names)
}
\arguments{
\item{data}{list of dataframes with original data}

\item{data_pca}{list of dataframes with principal component data}

\item{classes}{1-column dataframe with classes}

\item{names}{vector of names for each dataframe in list}
}
\description{
Plot correlation circle plots and biplots:
list of dataframes, list of pc, classes dataframe, names vector -> plots
}
