% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{plot_diablo}
\alias{plot_diablo}
\title{Make plots for diablo}
\usage{
plot_diablo(data, ncomp = 0, outdir = "./", data_names = NA, keepvar = "")
}
\arguments{
\item{data}{list of dataframes with quantitative multi-omics data.}

\item{ncomp}{integer assigning number of components for multi-block sPLSDA. Defaults to 0.}

\item{outdir}{string to outfile directory. Defaults to "./"}

\item{data_names}{names of individual omics data blocks. Defaults to NA.}

\item{keepvar}{name of kept variable. Defaults to "".}
}
\description{
Run the multi-block splsda on the multi omics data
}
