% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot PLSDA loadings}
\usage{
plot_loadings(
  data,
  contrib = "max",
  ncomp = 2,
  method = "median",
  ndisplay = 50,
  title = "Loadings"
)
}
\arguments{
\item{data}{list of dataframes with original data}

\item{contrib}{string to determine loading weights "max", "min". Defaults to "max".}

\item{ncomp}{integer assigning number of components for plot. Defaults to 2.}

\item{method}{string describing method of computation. Defaults to "median"}

\item{ndisplay}{integer number of features to display. Defaults to 50.}

\item{title}{string to name plots with. Defaults to "Loadings".}
}
\description{
Plot PLSDA loadings (weight of variable contribution to category)
}
