% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{plot_pca_single}
\alias{plot_pca_single}
\title{Plots PCA without performing multilevel decomposition}
\usage{
plot_pca_single(data, classes, pch = NA, title = "", ncomp = 0, show = FALSE)
}
\arguments{
\item{data}{dataframe with original data}

\item{classes}{1-column dataframe with classes}

\item{pch}{1-column dataframe with longitudinal measurement info. Defaults to NA.}

\item{title}{plot title. Defaults to ""}

\item{ncomp}{number of components to plot, if 0 estimate automatically. Defaults to 0.}

\item{show}{show the plot in the R graphics output, does not save. Defaults to FALSE.}
}
\description{
Plots PCA for non-repeated measurements
}
\seealso{
\code{\link[=plot_pca_single]{plot_pca_single()}}, \code{\link[=plot_pca_multilevel]{plot_pca_multilevel()}}
}
