% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{plot_plsda}
\alias{plot_plsda}
\title{Plot PLSDA output}
\usage{
plot_plsda(data, classes, pch, title = "", ncomp = 0)
}
\arguments{
\item{data}{list of dataframes with original data}

\item{classes}{1-column dataframe with classes}

\item{pch}{1-column dataframe with repeated measurments. Defaults to NA.}

\item{title}{string to name plots with. Defaults to "".}

\item{ncomp}{integer assigning number of components for PLSDA. Defaults to 0.}
}
\description{
Plot PLSDA output (PCA-like plots)
}
