% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{predict_diablo}
\alias{predict_diablo}
\title{Predict with weights}
\usage{
predict_diablo(data, test, classes)
}
\arguments{
\item{data}{dataframe with quantitative multi-omics data.}

\item{test}{dataframe with test data}

\item{classes}{1-column dataframe of classes}
}
\description{
Predict diablo with weights
}
\keyword{Internal}
