% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{remap_data}
\alias{remap_data}
\title{Map feature names to feature id}
\usage{
remap_data(data, mapping)
}
\arguments{
\item{data}{path to dataframe with data}

\item{mapping}{path to dataframe with mapping}
}
\description{
Remap feature names to feature id, format: dataframe, dataframe -> dataframe
Feature names can only reach a certain length and must be unique.
So you can use this if you want to remap identifiers out to something else.
}
\seealso{
\code{\link[=parse_mappings]{parse_mappings()}}
}
