% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{remove_na_class}
\alias{remove_na_class}
\title{Remove NA values in data by class}
\usage{
remove_na_class(data, classes, missing_as = NA)
}
\arguments{
\item{data}{dataframe with na values.}

\item{classes}{1-column dataframe with labelled samples matching order.}

\item{missing_as}{this value to represent missing values. Defaults to NA.}
}
\description{
Remove NA values from data and count: dataframe with na -> dataframe clean
If your NA values are represented as 0 or something else, convert them first.
}
\seealso{
\code{\link[=remove_na_class]{remove_na_class()}}, \code{\link[=remove_na_prop]{remove_na_prop()}}, \code{\link[=remove_novar]{remove_novar()}}
}
