% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{remove_na_prop}
\alias{remove_na_prop}
\title{Remove NA values in data by proportion}
\usage{
remove_na_prop(data, na_prop = 0.3, zero_as_na = TRUE)
}
\arguments{
\item{data}{list of dataframes with na values.}

\item{na_prop}{drop column if proportion of NA >= na_prop. Defaults to 0.3.}

\item{zero_as_na}{treat zero values as NA. Defaults to TRUE.}
}
\description{
Remove NA values from data and count: dataframe with na -> dataframe clean
If your NA values are represented as 0 or something else, convert them first.
}
\seealso{
\code{\link[=remove_na_class]{remove_na_class()}}, \code{\link[=remove_na_prop]{remove_na_prop()}}, \code{\link[=remove_novar]{remove_novar()}}
}
