% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{remove_novar}
\alias{remove_novar}
\title{Remove 0 variance columns in data}
\usage{
remove_novar(data)
}
\arguments{
\item{data}{dataframe with na values.}
}
\description{
Remove 0 variance columns in data: dataframe with na -> dataframe clean
}
\seealso{
\code{\link[=remove_na_class]{remove_na_class()}}, \code{\link[=remove_na_prop]{remove_na_prop()}}, \code{\link[=remove_novar]{remove_novar()}}
}
