% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace missing values}
\usage{
replace_missing(data, imputed)
}
\arguments{
\item{data}{dataframe with original data}

\item{imputed}{dataframe with imputed data}
}
\description{
Replace missing values: list of named dataframes with NA -> list of dataframes
This replaces missing values in the original dataframe with the corresponding imputed value
}
\seealso{
\code{\link[=impute_missing]{impute_missing()}}, \code{\link[=replace_missing]{replace_missing()}}, \code{\link[=cor_imputed_unimputed]{cor_imputed_unimputed()}}
}
