% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{show_na_prop}
\alias{show_na_prop}
\title{Show proportion of NA values in data}
\usage{
show_na_prop(data, name = "")
}
\arguments{
\item{data}{dataframe with na values.}

\item{name}{name the plot. Defaults to "".}
}
\description{
Show proportion of NA values in data: dataframe -> plot
If your NA values are represented as 0 or something else, convert them first.
}
