% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{show_variance}
\alias{show_variance}
\title{Show distribution of feature variance in data}
\usage{
show_variance(data, do_log2 = TRUE)
}
\arguments{
\item{data}{dataframe with na values.}

\item{do_log2}{perform log2 transform for visualisation. Defaults to TRUE.}
}
\description{
Show distribution of feature variance in data: dataframe -> plot of variances
}
\seealso{
\code{\link[=remove_na_class]{remove_na_class()}}, \code{\link[=remove_na_prop]{remove_na_prop()}}, \code{\link[=remove_novar]{remove_novar()}}
}
