% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{tune_diablo_keepx}
\alias{tune_diablo_keepx}
\title{Tune sPLSDA multi-block (DIABLO) features to keep}
\usage{
tune_diablo_keepx(
  data,
  classes,
  ncomp,
  design,
  test_keepX = c(5, 50, 100),
  cpus = 2,
  dist = "centroids.dist",
  progressBar = TRUE
)
}
\arguments{
\item{data}{list of dataframes with original data.}

\item{classes}{1-column dataframe with classes.}

\item{ncomp}{integer assigning number of components for multi-block sPLSDA. Defaults to 0.}

\item{design}{matrix with linkage}

\item{test_keepX}{vector of integers for feature selection. Defaults to c(5, 50, 100).}

\item{cpus}{integer number of cpus for parallel processing. Defaults to 2.}

\item{dist}{string describing distance metric "centroids.dist", "max.dist", "mahalanobis.dist". Defaults to "centroids.dist"}

\item{progressBar}{boolean showing progress bar. Defaults to TRUE.}
}
\description{
Tune sPLSDA multi-block features to keep
}
\seealso{
\code{\link[=create_design]{create_design()}}, \code{\link[=tune_diablo_ncomp]{tune_diablo_ncomp()}}, \code{\link[=tune_diablo_keepx]{tune_diablo_keepx()}}, \code{\link[=run_diablo]{run_diablo()}}
}
