% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{tune_diablo_ncomp}
\alias{tune_diablo_ncomp}
\title{Tune sPLSDA multi-block (DIABLO) number of components}
\usage{
tune_diablo_ncomp(data, classes, design, ncomp = 0, cpus = 1)
}
\arguments{
\item{data}{list of dataframes with original data.}

\item{classes}{1-column dataframe with classes.}

\item{design}{matrix with linkage}

\item{ncomp}{integer assigning number of components for multi-block sPLSDA. Defaults to 0.}

\item{cpus}{integer number of cpus for parallel processing. Defaults to 2.}
}
\description{
Tune sPLSDA multi-block optimal components
}
\seealso{
\code{\link[=create_design]{create_design()}}, \code{\link[=tune_diablo_ncomp]{tune_diablo_ncomp()}}, \code{\link[=tune_diablo_keepx]{tune_diablo_keepx()}}, \code{\link[=run_diablo]{run_diablo()}}
}
