% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{tune_splsda}
\alias{tune_splsda}
\title{Tune sPLSDA}
\usage{
tune_splsda(
  data,
  classes,
  names,
  multilevel = NULL,
  ncomp = 3,
  nrepeat = 10,
  logratio = "none",
  test_keepX = c(5, 50, 100),
  validation = "loo",
  folds = 10,
  dist = "centroids.dist",
  cpus = 2,
  progressBar = TRUE
)
}
\arguments{
\item{data}{list of dataframes with original data.}

\item{classes}{1-column dataframe with classes.}

\item{names}{vector of names for dataframes.}

\item{multilevel}{1-column dataframe with repeated measurments. Defaults to NULL.}

\item{ncomp}{integer assigning number of components for PLSDA. Defaults to 0.}

\item{nrepeat}{integer of repetitions for cross-validation. Defaults to 10.}

\item{logratio}{string to outfile directory. Defaults to "./"}

\item{test_keepX}{vector of integers for feature selection. Defaults to c(5, 50, 100).}

\item{validation}{string describing validation "loo", "M-fold". Defaults to "loo"}

\item{folds}{integer describing folds for M-fold validation only. Defaults to 10.}

\item{dist}{string describing distance metric "centroids.dist", "max.dist", "mahalanobis.dist". Defaults to "centroids.dist"}

\item{cpus}{integer number of cpus for parallel processing. Defaults to 2.}

\item{progressBar}{boolean showing progress bar. Defaults to TRUE.}
}
\description{
Tune sPLSDA optimal components
}
\seealso{
\code{\link[=classify_splsda]{classify_splsda()}}
}
