% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_omics_data.R
\docType{data}
\name{two_omics}
\alias{two_omics}
\title{COVID-19 patient study}
\format{
A list of two named dataframes containing two omics datasets. Also contains class information.
}
\usage{
data(two_omics)
}
\description{
Data from a COVID-19 patient study. Contains 100 anonymised patients.
Patients are divided into two categories, less severe and more severe.
In this subset, two omics datasets are featured - lipidome and metabolome.
The original dataset contains four omics datasets.
}
\examples{
data(two_omics)
}
\references{
Overmyer, K.A., Shishkova, E., Miller, I.J., Balnis, J., Bernstein, M.N., Peters-Clarke, T.M., Meyer, J.G., Quan, Q., Muehlbauer, L.K., Trujillo, E.A. and He, Y., 2020. Large-scale multi-omic analysis of COVID-19 severity. Cell systems.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/32743614}{PubMed})
}
\keyword{datasets}
