% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{zero_to_na}
\alias{zero_to_na}
\title{Turn zero into NA}
\usage{
zero_to_na(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Turn zero into NA: dataframe with 0 -> dataframe with NA
}
