% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_pipeline.R
\name{run_diablo}
\alias{run_diablo}
\title{Run the multi-omic pipeline}
\usage{
run_diablo(data, classes, ncomp, design, keepx = NULL)
}
\arguments{
\item{data}{list of dataframes with quantitative multi-omics data.}

\item{classes}{1-column dataframe with classes.}

\item{ncomp}{integer assigning number of components for multi-block sPLSDA.}

\item{design}{matrix with linkage.}

\item{keepx}{vector of integers for feature selection. Defaults to NULL.}
}
\description{
Run the multi-block splsda on the multi omics data
}
