% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPH2819.R, R/two_omics_data.R
\docType{data}
\name{two_omics}
\alias{two_omics}
\title{BPH2819 sepsis study}
\format{
A list of named dataframes containing class information and omics data.

A list of two named dataframes containing two omics datasets. Also contains class information.
}
\usage{
data(BPH2819)

data(two_omics)
}
\description{
Data from a sepsis study. 12 samples of Staphylococcus aureus strain BPH2819.
Divided into two categories, bacteria grown on RPMI media and from human sera.
Three omics datasets are featured - metabolome, proteome, translatome.
Original data had a small amount of missing values.
Missing values were filled in with data imputation.
For convenience and reproducibility this data object contains the imputed data.
Original unimputed data is available for reference.

Data from a COVID-19 patient study. Contains 100 anonymised patients.
Patients are divided into two categories, less severe and more severe.
In this subset, two omics datasets are featured - lipidome and metabolome.
The original dataset contains four omics datasets.
}
\examples{
data(BPH2819)
data(two_omics)
}
\references{
Mu, A., Klare, W.P., Baines, S.L. et al. Integrative omics identifies conserved and pathogen-specific responses of sepsis-causing bacteria. Nat Commun 14, 1530 (2023). https://doi.org/10.1038/s41467-023-37200-w
(\href{https://doi.org/10.1038/s41467-023-37200-w}{DOI})

Overmyer, K.A., Shishkova, E., Miller, I.J., Balnis, J., Bernstein, M.N., Peters-Clarke, T.M., Meyer, J.G., Quan, Q., Muehlbauer, L.K., Trujillo, E.A. and He, Y., 2020. Large-scale multi-omic analysis of COVID-19 severity. Cell systems.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/32743614}{PubMed})
}
\keyword{datasets}
